+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2025-12-02
+
+### Added
+
+- "Modbus - Read Device Information" ability
+- Github Action to build payloads
+
+
 ## [1.3.0] - 2025-08-05
 
 ### Added
@@ -42,7 +50,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   [MITRE Caldera™ for OT](https://github.com/mitre/caldera-ot). Provides 10
   unique Caldera adversary emulation abilities specific to the Modbus protocol.
 
-[unreleased]: https://github.com/mitre/modbus/compare/v1.3.0...HEAD
+[unreleased]: https://github.com/mitre/modbus/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/mitre/modbus/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/mitre/modbus/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/mitre/modbus/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/mitre/modbus/compare/v1.0.0...v1.1.0